# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexo chargeback API",
    "version": "0.1.0",
    "description": "Synthetic transactions and a simulated card-network receiver. Currency amounts use exact integer strings."
  },
  "paths": {
    "/api/transactions": {
      "get": {
        "summary": "List owned imported transactions",
        "description": "Requires customer token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      },
                      "maxItems": 100
                    }
                  },
                  "required": [
                    "transactions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/chargebacks": {
      "get": {
        "summary": "List latest owned chargebacks",
        "description": "Requires customer token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "chargebacks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chargeback"
                      },
                      "maxItems": 100
                    }
                  },
                  "required": [
                    "chargebacks"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or replay a chargeback",
        "description": "Requires customer token. At most 4096 body bytes. Customer, network and currency derive from the owned imported transaction. Same customer/key and same canonical request returns the original chargeback; changed content or a second dispute for the same transaction conflicts. Receipt means durable intake, not network submission. Commit uncertainty returns 503; retry using the identical key and body.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chargeback"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "Created after durable commit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chargeback"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Owned transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict or transaction already disputed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[!-~]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChargeback"
              }
            }
          }
        }
      }
    },
    "/api/chargebacks/{id}": {
      "get": {
        "summary": "Get an owned chargeback",
        "description": "Requires customer token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chargeback"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/ops/summary": {
      "get": {
        "summary": "Read lifecycle counts and oldest pending age",
        "description": "Requires ops token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/ops/batches": {
      "get": {
        "summary": "List latest batches",
        "description": "Requires ops token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "batches": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Batch"
                      },
                      "maxItems": 100
                    }
                  },
                  "required": [
                    "batches"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/ops/batches/{id}": {
      "get": {
        "summary": "Inspect a batch and its latest attempts and audit events",
        "description": "Requires ops token. Collection results are bounded to 100 rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "batch": {
                      "$ref": "#/components/schemas/Batch"
                    },
                    "attempts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Attempt"
                      },
                      "maxItems": 100
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditEvent"
                      },
                      "maxItems": 100
                    }
                  },
                  "required": [
                    "batch",
                    "attempts",
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry with the same idempotency key for a write",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Provisioned opaque token, SHA-256 hash stored durably; supports expiry and revocation."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "CreateChargeback": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount_minor": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "Exact positive minor units, at most signed 64-bit integer maximum; never a JSON number."
          },
          "reason": {
            "type": "string",
            "enum": [
              "FRAUD",
              "DUPLICATE",
              "NOT_RECEIVED",
              "OTHER"
            ]
          }
        },
        "required": [
          "transaction_id",
          "amount_minor",
          "reason"
        ],
        "additionalProperties": false
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "network": {
            "type": "string",
            "enum": [
              "visa",
              "mastercard"
            ]
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "amount_minor": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "Exact positive minor units, at most signed 64-bit integer maximum; never a JSON number."
          },
          "disputed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "network",
          "currency",
          "amount_minor",
          "disputed"
        ],
        "additionalProperties": false
      },
      "Chargeback": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount_minor": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "Exact positive minor units, at most signed 64-bit integer maximum; never a JSON number."
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "network": {
            "type": "string",
            "enum": [
              "visa",
              "mastercard"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "FRAUD",
              "DUPLICATE",
              "NOT_RECEIVED",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "building",
              "ready",
              "unknown",
              "submitted",
              "acknowledged",
              "rejected",
              "conflict"
            ]
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "received_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "transaction_id",
          "amount_minor",
          "currency",
          "network",
          "reason",
          "status",
          "batch_id",
          "received_at"
        ],
        "additionalProperties": false
      },
      "Batch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slot": {
            "type": "string",
            "format": "date-time"
          },
          "network": {
            "type": "string",
            "enum": [
              "visa",
              "mastercard"
            ]
          },
          "partition": {
            "type": "integer",
            "minimum": 0
          },
          "state": {
            "type": "string",
            "enum": [
              "building",
              "ready",
              "unknown",
              "submitted",
              "acknowledged",
              "rejected",
              "conflict"
            ]
          },
          "row_count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "byte_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "publication_started": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "next_attempt_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ack": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "slot",
          "network",
          "partition",
          "state",
          "row_count",
          "object_key",
          "sha256",
          "byte_count",
          "publication_started",
          "created_at",
          "updated_at",
          "next_attempt_at",
          "last_error",
          "ack"
        ],
        "additionalProperties": false
      },
...
      "Summary": {
        "type": "object",
        "properties": {
          "chargebacks": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "batches": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "oldest_pending_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "chargebacks",
          "batches",
          "oldest_pending_seconds"
        ],
        "additionalProperties": false
      }
    }
  }
}